requirements.txt
aiohttp
chardet
multidict
aiopvapi
pyserial
requests
wheel
telepot
async_timeout
python-slugify